Change Success/Error blink intervals
In practice with the HOTP checks, we would like the success interval to
be longer as a user will first be looking at their screen and may miss
the success blinking on the device. Doubling the interval should be
enough.

In the case of an error, however, we should treat it differently and
blink MUCH longer to make sure the user notices the blinking red LED.
Blinking red indefinitely would be ideal but instead I just increased
the interval to a much larger amount.