Introduce debug build with BUILD_DEBUG=1 switch
Removing unused code sections results in a linking error, hence part of the code has to be discarded.
To be confirmed, whether this changes the runtime behavior for the release version, and enable there as well.
GC of the unused code is required for the development build, which does not fit the flash size otherwise.

Mentioned error message:
/home/sz/work/software/gcc-arm-old/gcc-arm-none-eabi-8-2018-q4-major/bin/../lib/gcc/arm-none-eabi/8.2.1/../../../../ar
m-none-eabi/bin/ld: section .ARM.exidx LMA [000000000800b9cc,000000000800b9d3] overlaps section .data LMA [00000000080
0b9cc,000000000800c02b]